+++ builtins.c
@@ -6,80 +6,52 @@
 /*   By: aalsuwai <aalsuwai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/02 12:12:03 by aalsuwai          #+#    #+#             */
-/*   Updated: 2022/04/03 17:34:30 by aalsuwai         ###   ########.fr       */
+/*   Updated: 2022/04/04 15:18:34 by aalsuwai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-int	env(t_parser_info *p)
+int	builtin_execute(t_parser_info *p, int i)
 {
-	int	i;
-
-	i = 0;
-	while (p->env[i])
-		printf("%s\n", p->env[i++]);
-	return(0);
-}
-
-int	pwd(t_parser_info *p)
-{
-	char	cwd[PATH_MAX];
-	if (!getcwd(cwd, sizeof(cwd)))
+	if (compare_caseless(p->cmd[i][0], "echo"))
+		return (echo(p, p->cmd[i]));
+	else if (ft_smartncmp(p->cmd[i][0], "cd", 3))
+		return (cd(p, p->cmd[i]));
+	else if (compare_caseless(p->cmd[i][0], "pwd"))
+		return (pwd(p));
+	else if (ft_smartncmp(p->cmd[i][0], "export", 7))
+		return (export(p, p->cmd[i]));
+	else if (ft_smartncmp(p->cmd[i][0], "unset", 6))
+		return (unset(p, p->cmd[i]));
+	else if (compare_caseless(p->cmd[i][0], "env"))
+		return (env(p));
+	else if (ft_smartncmp(p->cmd[i][0], "clear", 6))
+		return (clear(p));
+	else if (ft_smartncmp(p->cmd[i][0], "exit", 5))
 	{
-		perror("babyshell: pwd");
-		p->exit_code = 1;
+		baby_exit(p, p->cmd[i]);
+		return (0);
 	}
-	else
-		printf("%s\n", cwd);
-	p->exit_code = 0;
-	return (0);
+	return (-1);
 }
 
-int	echo_index(char **argv)
+int	builtin_check(t_parser_info *p, int i)
 {
-	int	index;
-	int	array_index;
-
-	array_index = 1;
-	while (argv[array_index])
-	{
-		if (argv[array_index][0] != '-' || argv[array_index][1] != 'n')
-			return (array_index);
-		index = 2;
-		while (argv[array_index][index])
-		{
-			if (argv[array_index][index] != 'n')
-				return (array_index);
-			index++;
-		}
-		array_index++;
-	}
-	return (array_index);
+	if (ft_smartncmp(p->cmd[i][0], "cd", 3) || \
+	ft_smartncmp(p->cmd[i][0], "export", 7) || \
+	ft_smartncmp(p->cmd[i][0], "unset", 6) || \
+	ft_smartncmp(p->cmd[i][0], "clear", 6) || \
+	ft_smartncmp(p->cmd[i][0], "exit", 5))
+		return (0);
+	else if (compare_caseless(p->cmd[i][0], "echo") || \
+	compare_caseless(p->cmd[i][0], "pwd") || \
+	compare_caseless(p->cmd[i][0], "env"))
+		return (1);
+	return (2);
 }
 
-int	echo(t_parser_info *p, char **argv)
-{
-	int		i;
-	bool	newline_flag;
-
-	i = echo_index(argv);
-	newline_flag = true;
-	if (i > 1)
-		newline_flag = false;
-	while (argv[i])
-	{
-		printf("%s", argv[i]);
-		if (argv[++i])
-			printf(" ");
-	}
-	if (newline_flag)
-		printf("\n");
-	p->exit_code = 0;
-	return (0);
-}
-
-void	lonely_export(t_parser_info *p)
+static void	lonely_export(t_parser_info *p)
 {
 	int	i;
 	int	env_index;
@@ -103,7 +75,7 @@ int	export(t_parser_info *p, char **cmd)
 {
 	int		i;
 	bool	failed;
-	
+
 	i = 1;
 	failed = false;
 	if (!cmd[1])
@@ -116,7 +88,7 @@ int	export(t_parser_info *p, char **cmd)
 	{
 		p->env = export_env(p, p->env, cmd[i++]);
 		if (p->exit_code == 1)
-			failed = true;	
+			failed = true;
 	}
 	if (failed)
 		p->exit_code = 1;
@@ -127,7 +99,7 @@ int	unset(t_parser_info *p, char **cmd)
 {
 	int		i;
 	bool	failed;
-	
+
 	i = 1;
 	failed = false;
 	while (cmd[i])
@@ -139,95 +111,4 @@ int	unset(t_parser_info *p, char **cmd)
 	if (failed)
 		p->exit_code = 1;
 	return (0);
-}
-
-int	cd(t_parser_info *p, char **argv)
-{
-	char	*temp;
-	char	*save;
-	char	cwd[PATH_MAX];
-	
-	p->exit_code = 0;
-	getcwd(cwd, sizeof(cwd));
-	save = ft_strdup(cwd);
-	ft_bzero(cwd, PATH_MAX);
-	if (!argv[1])
-	{
-		temp = local_getenv("HOME", p->env);
-		if (!temp)
-		{
-			printf("babyshell: cd: HOME not set\n");
-			p->exit_code = 1;
-
-		}
-		else if (chdir(temp) == -1)
-		{
-			temp = ft_strjoin("babyshell: cd: ", temp);
-			perror(temp);
-			free(temp);
-			p->exit_code = 1;
-		}
-	}
-	else
-	{
-		if (chdir(argv[1]) == -1)
-		{
-			temp = ft_strjoin("babyshell: cd: ", argv[1]);
-			perror(temp);
-			free(temp);
-			p->exit_code = 1;
-		}
-	}
-	if (p->exit_code == 0)//if cd succeeds
-	{
-		if (!p->oldpwd_dont_update)
-		{
-			temp = ft_strjoin("OLDPWD=", save);
-			p->env = export_env(p, p->env, temp);
-			free(temp);
-		}
-		getcwd(cwd, sizeof(cwd));
-		if (local_getenv("PWD", p->env))
-		{
-			temp = ft_strjoin("PWD=", cwd);
-			p->env = export_env(p, p->env, temp);
-			free(temp);
-		}
-	}
-	free(save);
-	return(0);
-}
-
-void	baby_exit(t_parser_info *p, char **cmd)
-{
-	bool	exit_flag;
-
-	exit_flag = true;
-	if (!p->pipes_count)
-		printf("exit\n");
-	if (!cmd[1])
-		p->exit_code = (unsigned char)p->exit_code;
-	else
-	{
-		if (ft_str_isdigit(cmd[1]) == 1 && !cmd[2] && !check_longmax(cmd[1]))
-			p->exit_code = (unsigned char)ft_atoi(cmd[1]);
-		else if (ft_str_isdigit(cmd[1]) == 1 && cmd[2] && !check_longmax(cmd[1]))
-		{
-			p->exit_code = 1;
-			exit_flag = false;
-			printf("babyshell: exit: too many arguments\n");
-		}
-		else
-		{
-			p->exit_code = 255;
-			printf("babyshell: exit: %s: numeric argument required\n", cmd[1]);
-		}
-	}
-	if (exit_flag && !p->pipes_count)
-	{
-		free_everything(p);
-		free_double_char(p->env);
-		free(p->cmd_path);
-		exit(p->exit_code);
-	}
 }
