+++ get_cmd_path.c
@@ -6,7 +6,7 @@
 /*   By: aalsuwai <aalsuwai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/14 16:55:09 by anasr             #+#    #+#             */
-/*   Updated: 2022/04/03 17:34:45 by aalsuwai         ###   ########.fr       */
+/*   Updated: 2022/04/04 14:01:39 by aalsuwai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,15 +48,13 @@ void	change_cmd(t_parser_info *p, int array_i)
 	}
 	else
 	{
-		// if (!ft_strncmp(ft_strrchr(p->cmd[array_i][0], '/') + 1, "minishell", 10))
-		// 	nested_minishell(p);
 		p->cmd_path = ft_strdup(p->cmd[array_i][0]);
 		free(p->cmd[array_i][0]);
 		p->cmd[array_i][0] = ft_strdup(ft_strrchr(p->cmd_path, '/') + 1);
 	}
 }
 
-char	*join_cmd_to_path(char *path, char *cmd)
+static char	*join_cmd_to_path(char *path, char *cmd)
 {
 	char	*temp;
 	char	*joined;
@@ -67,7 +65,14 @@ char	*join_cmd_to_path(char *path, char *cmd)
 	return (joined);
 }
 
-/*Malloc occurs here*/
+static void	perror_n_exitcode(char *cmd, t_parser_info *p, bool no_cmd)
+{
+	p->exit_code = 127;
+	if (no_cmd == false)
+		printf("babyshell: %s: No such file or directory\n", cmd);
+	else
+		printf("babyshell: %s:  command not found\n", cmd);
+}
 
 char	*get_cmd_path(char *cmd, t_parser_info *p)
 {
@@ -81,8 +86,7 @@ char	*get_cmd_path(char *cmd, t_parser_info *p)
 	temp_path = local_getenv("PATH", p->env);
 	if (!temp_path)
 	{
-		printf("babyshell: %s: No such file or directory\n", cmd);
-		p->exit_code = 127;
+		perror_n_exitcode(cmd, p, false);
 		return (0);
 	}
 	paths_array = ft_split(temp_path, ':');
@@ -94,9 +98,7 @@ char	*get_cmd_path(char *cmd, t_parser_info *p)
 		free(temp_path);
 		i++;
 	}
-	printf("babyshell: %s:  command not found\n", cmd);
+	perror_n_exitcode(cmd, p, true);
 	free_double_char(paths_array);
-	p->exit_code = 127;
 	return (NULL);
-	// exit (127); // error message
 }
