+++ local_getenv.c
@@ -6,7 +6,7 @@
 /*   By: aalsuwai <aalsuwai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/14 15:50:10 by aalsuwai          #+#    #+#             */
-/*   Updated: 2022/03/21 11:35:08 by aalsuwai         ###   ########.fr       */
+/*   Updated: 2022/04/04 14:02:35 by aalsuwai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,10 +28,10 @@ static int	cmp_till_equal(char *p_env, char *var)
 	return (1);
 }
 
-char	*local_getenv(char *var, char **p_env) // local_getenv("NAME", p->env)
+char	*local_getenv(char *var, char **p_env)
 {
 	int	i;
-	
+
 	i = 0;
 	while (p_env[i])
 	{
@@ -41,13 +41,3 @@ char	*local_getenv(char *var, char **p_env) // local_getenv("NAME", p->env)
 	}
 	return (0);
 }
-
-// int	main(int argc, char **argv, char **env)
-// {
-// 	(void)argc;
-// 	(void)argv;
-// 	printf("%stest 1:%s%s %s%s\n", GREEN2, RESET, GREEN1, local_getenv("USER", env), RESET);
-// 	printf("%stest 2:%s%s %s%s\n", YELLOW2, RESET, YELLOW1, local_getenv("USERR", env), RESET);
-// 	printf("%stest 3:%s%s %s%s\n", LIGHT_BLUE2, RESET, LIGHT_BLUE1, local_getenv("PATH", env), RESET);
-// 	printf("%stest 4:%s%s %s%s\n", PURPLE2, RESET, PURPLE1, local_getenv("OWO", env), RESET);
-// }
